-dashboard ui create - perform crud for user

+++ lib/ui/home/home_controller.dart
@@ -6,6 +6,7 @@ import '../../models/user/create_user_model.dart';
 import '../../models/user/get_all_user.dart';
 import '../../network/repo/user_repo.dart';
 import '../../utils/strings.dart';
+
 import 'home_model.dart';
 
 class HomeController extends GetxController {
@@ -24,36 +25,41 @@ class HomeController extends GetxController {
   }
 
   Future<void> deleteUser(num id) async {
-    startLoading();
-    BaseModel response = await repo.deleteUser(uid: id);
-    if (response.body != null) {
-      var data = (response.body as NeoBaseModel);
-      if (data.status ?? false) {
-        getAlluser(model.currentPage);
+    if(model.isLoggedIn()){
+      startLoading();
+      BaseModel response = await repo.deleteUser(uid: id);
+      if (response.body != null) {
+        var data = (response.body as NeoBaseModel);
+        if (data.status ?? false) {
+          getAlluser(model.currentPage);
+        }
       }
     }
   }
 
   Future<void> editUser(UserData user) async {
-    BaseModel response = await repo.updateUser(
-        uid: user.uid,
-        email: model.emailController.text,
-        firstName: model.firstNameController.text,
-        lastName: model.lastNameController.text,
-        passWord: model.passwordController.text,
-        phoneNumber: model.mobileNumberController.text);
-    if (response.body != null) {
-      var data = (response.body as CreateUserModel);
-      if (data.status ?? false) {
-        clearTextFields();
-        stopLoading();
-        Get.back();
-        getAlluser(model.currentPage);
+    if (isEditUserValidate() && model.isLoggedIn()) {
+      BaseModel response = await repo.updateUser(
+          uid: user.uid,
+          email: model.emailController.text,
+          firstName: model.firstNameController.text,
+          lastName: model.lastNameController.text,
+          passWord: model.passwordController.text,
+          phoneNumber: model.mobileNumberController.text);
+      if (response.body != null) {
+        var data = (response.body as CreateUserModel);
+        if (data.status ?? false) {
+          clearTextFields();
+          stopLoading();
+          Get.back();
+          getAlluser(model.currentPage);
+        } else {
+          data.msg.errorSnack();
+        }
       } else {
-        data.msg.errorSnack();
+        stopLoading();
+        someThingWentWrong.errorSnack();
       }
-    } else {
-      someThingWentWrong.errorSnack();
     }
   }
 
@@ -63,51 +69,66 @@ class HomeController extends GetxController {
     model.emailController.clear();
     model.mobileNumberController.clear();
     model.passwordController.clear();
+    model.setPasswordError(null);
+    model.setMobileNumberError(null);
+    model.setLastNameError(null);
+    model.setFirstNameError(null);
+    model.setEmailError(null);
   }
 
   Future<void> addUser() async {
-    startLoading();
-    BaseModel response = await repo.createUser(
-        email: model.emailController.text,
-        firstName: model.firstNameController.text,
-        lastName: model.lastNameController.text,
-        passWord: model.passwordController.text,
-        phoneNumber: model.mobileNumberController.text);
-    if (response.body != null) {
-      var data = response.body as CreateUserModel;
-      if (data.status ?? false) {
-        clearTextFields();
-        stopLoading();
-        Get.back();
-        getAlluser(model.currentPage);
+    if (isAddUserValidate() && model.isLoggedIn()) {
+      startLoading();
+      BaseModel response = await repo.createUser(
+          email: model.emailController.text,
+          firstName: model.firstNameController.text,
+          lastName: model.lastNameController.text,
+          passWord: model.passwordController.text,
+          phoneNumber: model.mobileNumberController.text);
+      if (response.body != null) {
+        var data = response.body as CreateUserModel;
+        if (data.status ?? false) {
+          clearTextFields();
+          stopLoading();
+          Get.back();
+          getAlluser(model.currentPage);
+        } else {
+          data.msg.errorSnack();
+          stopLoading();
+        }
       } else {
-        data.msg.errorSnack();
+        someThingWentWrong.errorSnack();
+        stopLoading();
       }
-    } else {
-      someThingWentWrong.errorSnack();
     }
   }
 
   Future<void> getAlluser(int currentPage) async {
-    startLoading();
-    BaseModel response = await repo.getUserList(currentPage);
-    if (response.body != null) {
-      var data = response.body as AllUserModel;
-      if (data.status ?? false) {
-        model.userList.clear();
-        model.addToUserList(data.data ?? []);
-        if(data.data?.length!=5){
-          model.setStopPagination(true);
-        }else{
-          model.setStopPagination(false);
+    if(model.isLoggedIn()){
+      startLoading();
+      BaseModel response = await repo.getUserList(
+        currentPage,
+        model.currentSelectedValue!,
+        model.userSearchController.text,
+      );
+      if (response.body != null) {
+        var data = response.body as AllUserModel;
+        if (data.status ?? false) {
+          model.userList.clear();
+          model.addToUserList(data.data ?? []);
+          if (data.data?.length != model.currentSelectedValue) {
+            model.setStopPagination(true);
+          } else {
+            model.setStopPagination(false);
+          }
+        } else {
+          data.msg.errorSnack();
         }
       } else {
-        data.msg.errorSnack();
+        someThingWentWrong.errorSnack();
       }
-    } else {
-      someThingWentWrong.errorSnack();
+      stopLoading();
     }
-    stopLoading();
   }
 
   startLoading() {
@@ -121,18 +142,123 @@ class HomeController extends GetxController {
   }
 
   void previousPage() {
-    if(model.currentPage>1){
-      model.setCurrentPage(model.currentPage-1);
+    if (model.currentPage > 1) {
+      model.setCurrentPage(model.currentPage - 1);
       getAlluser(model.currentPage);
     }
   }
 
   void nextPage() {
-    if(!model.stopPagination){
-      model.setCurrentPage(model.currentPage+1);
+    if (!model.stopPagination) {
+      model.setCurrentPage(model.currentPage + 1);
       getAlluser(model.currentPage);
     }
   }
 
-  void logOut() {}
+
+
+  void changeSelection(int? newValue) {
+    model.setCurrentSelectedValue(newValue);
+    model.setCurrentPage(1);
+    getAlluser(model.currentPage);
+  }
+
+  bool isAddUserValidate() {
+    if (model.emailController.text.isEmpty ||
+        model.firstNameController.text.isEmpty ||
+        model.lastNameController.text.isEmpty ||
+        model.mobileNumberController.text.isEmpty ||
+        model.passwordController.text.isEmpty ||
+        !model.emailController.text.isEmail) {
+      if (model.emailController.text.isEmpty) {
+        model.setEmailError("Please Enter Email");
+      } else {
+        if (!model.emailController.text.isEmail) {
+          model.setEmailError("Please Enter Correct Email");
+        } else {
+          model.setEmailError(null);
+        }
+      }
+      if (model.firstNameController.text.isEmpty) {
+        model.setFirstNameError("Please Enter FirstName");
+      } else {
+        model.setFirstNameError(null);
+      }
+      if (model.lastNameController.text.isEmpty) {
+        model.setLastNameError("Please Enter LastName");
+      } else {
+        model.setLastNameError(null);
+      }
+      if (model.mobileNumberController.text.isEmpty) {
+        model.setMobileNumberError("Please Enter mobile Number");
+      } else {
+        model.setMobileNumberError(null);
+      }
+      if (model.passwordController.text.isEmpty) {
+        model.setPasswordError("Please Enter Password");
+      } else {
+        model.setPasswordError(null);
+      }
+      update(['DialogBuilder']);
+      return false;
+    }
+    model.setPasswordError(null);
+    model.setMobileNumberError(null);
+    model.setLastNameError(null);
+    model.setFirstNameError(null);
+    model.setEmailError(null);
+    update(['DialogBuilder']);
+    return true;
+  }
+
+  bool isEditUserValidate() {
+    if (model.emailController.text.isEmpty ||
+        model.firstNameController.text.isEmpty ||
+        model.lastNameController.text.isEmpty ||
+        model.mobileNumberController.text.isEmpty ||
+        !model.emailController.text.isEmail) {
+      if (model.emailController.text.isEmpty) {
+        model.setEmailError("Please Enter Email");
+      } else {
+        if (!model.emailController.text.isEmail) {
+          model.setEmailError("Please Enter Correct Email");
+        } else {
+          model.setEmailError(null);
+        }
+      }
+      if (model.firstNameController.text.isEmpty) {
+        model.setFirstNameError("Please Enter FirstName");
+      } else {
+        model.setFirstNameError(null);
+      }
+      if (model.lastNameController.text.isEmpty) {
+        model.setLastNameError("Please Enter LastName");
+      } else {
+        model.setLastNameError(null);
+      }
+      if (model.mobileNumberController.text.isEmpty) {
+        model.setMobileNumberError("Please Enter mobile Number");
+      } else {
+        model.setMobileNumberError(null);
+      }
+      update(['DialogBuilder']);
+      return false;
+    }
+    model.setMobileNumberError(null);
+    model.setLastNameError(null);
+    model.setFirstNameError(null);
+    model.setEmailError(null);
+    update(['DialogBuilder']);
+    return true;
+  }
+
+  void searchUser() {
+    model.setCurrentPage(1);
+    getAlluser(model.currentPage);
+  }
+
+  void clearSearch() {
+    model.userSearchController.clear();
+    searchUser();
+  }
 }
